tests/wpt/web-platform-tests/html/canvas/offscreen/pixel-manipulation/2d.imageData.object.undefined.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.imageData.object.undefined</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.imageData.object.undefined</h1>
<p class="desc">ImageData.data converts undefined to 0</p>


<script>
var t = async_test("ImageData.data converts undefined to 0");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

var offscreenCanvas = new OffscreenCanvas(100, 50);
var ctx = offscreenCanvas.getContext('2d');

var imgdata = ctx.getImageData(0, 0, 10, 10);
imgdata.data[0] = 100;
imgdata.data[0] = undefined;
_assertSame(imgdata.data[0], 0, "imgdata.data[\""+(0)+"\"]", "0");
t.done();

});
</script>
